Extract ListUsersToolbarView to own file
